+++ packages/doars/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG.md
 
+## 1.1.1
+
+- Fix reactivity of $inContext context.
+
 ## 1.1.0
 
 - Added ignore directive.